rename mappe MånedÅrInput -> MånedÅrVelger
