+++ src/frontend/sider/Fagsak/Behandling/Sider/Behandlingsresultat/Eøs/EøsKomponenter/EøsPeriodeSkjema.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components';
 import { Fieldset } from '@navikt/ds-react';
 import type { Felt } from '@navikt/familie-skjema';
 
-import MånedÅrVelger from '../../../../../../../komponenter/MånedÅrInput/MånedÅrVelger';
+import MånedÅrVelger from '../../../../../../../komponenter/MånedÅrVelger/MånedÅrVelger';
 import type { IIsoMånedPeriode } from '../../../../../../../utils/dato';
 
 const FlexDiv = styled.div`
